test refactor - extract common FeedItemEventListener test helper.
